basic operator support + broadcast support + more below
this commit also converted the default number type in variables from ints to doubles.
it also adds the go to xy position block.

+++ src/scratch/ScratchSprite.hpp
@@ -20,7 +20,7 @@ public:
     ScratchSprite(ScratchStage& st);
 
     bool visible;
-    int x, y, size, direction;
+    double x, y, size, direction;
     bool draggable;
     RotationStyle rotationStyle;
 };
